Contracts compile + tests green (via WSL); fix 16 build errors
Built with cargo-odra 0.1.7 in WSL Ubuntu (no MSVC needed). Produces 5 distinct
deployable wasm; `cargo odra test` = 5 passed, 0 failed.

- add build.rs (odra_build::build) — was missing; root cause of byte-identical
  wasm and the build_schema link error during tests.
- reputation_registry: i128 -> i64 (Casper CLType has no i128).
- rwa_vault: `use odra::ContractRef` so cross-contract refs resolve.
- tests/integration: RwaVaultHostRef import, contract_address(), mut host refs.
- docs: drop invalid `-b casper` flag (cargo-odra 0.1.7), document WSL build env
  + optional wasm-opt (binaryen) caveat.

+++ contracts/src/reputation_registry.rs
@@ -4,10 +4,10 @@ use odra::prelude::*;
 
 #[odra::module(errors = Error)]
 pub struct ReputationRegistry {
-    // ponytail: verify against odra 2.8 — i128 must be a supported CL/odra
-    // storage type. If the macro rejects i128, fall back to i64 (scores are
-    // small) or a {sign:bool, mag:U256} odra_type. Spec asks for i128.
-    score: Mapping<Address, i128>,
+    // ponytail: spec asked for i128, but Casper CLType only goes up to i64
+    // (no i128 CLTyped/ToBytes impl). i64 holds reputation scores with room to
+    // spare; widen to a {sign:bool, mag:U256} odra_type only if it can overflow.
+    score: Mapping<Address, i64>,
     consumed_payment_proofs: Mapping<[u8; 32], bool>,
 }
 
@@ -25,12 +25,12 @@ impl ReputationRegistry {
 
     /// Adjust a score by a signed delta. `outcome_ref` links to the off-chain
     /// outcome that justified the adjustment (kept in the deploy/audit trail).
-    pub fn adjust(&mut self, addr: Address, delta: i128, _outcome_ref: [u8; 32]) {
+    pub fn adjust(&mut self, addr: Address, delta: i64, _outcome_ref: [u8; 32]) {
         let s = self.score.get_or_default(&addr);
         self.score.set(&addr, s + delta);
     }
 
-    pub fn score_of(&self, addr: Address) -> i128 {
+    pub fn score_of(&self, addr: Address) -> i64 {
         self.score.get_or_default(&addr)
     }
 }
